Production_Ready/rag_production/pyproject.toml
[project]
name = "rag-production"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=0.6.3",
    "docx2txt>=0.9",
    "faiss-cpu>=1.11.0",
    "fastapi>=0.115.12",
    "langchain>=0.3.25",
    "langchain-anthropic>=0.3.15",
    "langchain-community>=0.3.25",
    "langchain-core>=0.3.65",
    "langchain-google-genai>=2.1.5",
    "langchain-openai>=0.3.22",
    "langgraph>=0.4.8",
    "pydantic>=2.11.5",
    "pypdf>=5.6.0",
    "python-docx>=1.1.2",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.3",
]
